CPM-195: Move hooks
