+++ front-packages/akeneo-design-system/src/hooks/useDrop.ts
@@ -1,6 +1,5 @@
-import {DragEvent, useContext} from 'react';
-import {TableInputContext} from '../TableInputContext';
-import {useId} from '../../../../hooks';
+import {useId} from './index';
+import {DragEvent} from 'react';
 
 /**
  * Recursively find the draggable parent not to know which element got dropped on.
@@ -26,19 +25,22 @@ const generateReorderedIndices = (size: number, draggedIndex: number, droppedInd
   return arrayWithoutDraggedItem;
 };
 
-const useDrop = (tableSize: number, draggedElementIndex: number | null) => {
+const useDrop = (
+  tableSize: number,
+  draggedElementIndex: number | null,
+  onReorder: ((reorderedIndices: number[]) => void) | undefined
+) => {
   const tableId = useId('table_');
-  const {onReorder} = useContext(TableInputContext);
 
   const handleDrop = (event: DragEvent<HTMLTableSectionElement>) => {
+    event.stopPropagation();
+    event.preventDefault();
+
     if (event.currentTarget.dataset.tableId === tableId && onReorder && null !== draggedElementIndex) {
       const droppedElementIndex = getDropRow(event.target as HTMLElement);
       const newIndices = generateReorderedIndices(tableSize, draggedElementIndex, droppedElementIndex);
 
       onReorder(newIndices);
-
-      event.stopPropagation();
-      event.preventDefault();
     }
   };
