+++ front-packages/akeneo-design-system/src/hooks/usePlaceholderPosition.ts
